several corrections, full example/test

+++ inc/graph_construction.h
@@ -1,6 +1,6 @@
 #ifndef GRAPH_CONSTRUCT
 #define GRAPH_CONSTRUCT
 
-Graph get_2d_toric_code(int64_t lsize);
+Graph get_2d_toric_code(int lsize);
 
 #endif
